+++ src/components/Input/index.ts
@@ -1,2 +1,3 @@
 export * from './Input'
 export { CurrencyInput } from './CurrencyInput'
+export { SearchInput } from './SearchInput'